+++ src/image_io/fluff.rs
@@ -148,6 +148,8 @@ pub fn load_fluff64(reader: &mut impl Read) -> Result<VicImage, Error> {
         })
         .collect::<Result<Vec<vic::Char>, Error>>()?;
     let mut image = VicImage::with_content(ImgVec::new(video_buffer, width, height));
-    image.colors = GlobalColors([header.background, header.border, header.aux]);
+    image.set_global_color(GlobalColors::BACKGROUND, header.background);
+    image.set_global_color(GlobalColors::BORDER, header.border);
+    image.set_global_color(GlobalColors::AUX, header.aux);
     Ok(image)
 }